+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2020-01-12
+
+### rax-portal [1.0.1]
+- Fix: not re-render when the container changes
+
 ## 2020-12-18
 
 ### rax-image [2.2.4]